--- app/clients/__init__.py
@@ -1,1 +1,0 @@
-"""Clients for external data sources and storage backends."""
