+++ backend/jobs/update_performance_snapshot.py
@@ -2,19 +2,20 @@ import os
 import sys
 import warnings
 
-import fundamentus as fd
 import numpy as np
 import pandas as pd
+
+from shared.clients.fundamentus_client import get_resultado_raw, list_papel_setor
+from shared.config import PERFORMANCE_CSV_PATH
+
 
 warnings.filterwarnings("ignore")
 
 
 def carregar_estado_anterior():
-    """Carrega o estado anterior da carteira, se existir."""
-    if os.path.exists("data/performance.csv"):
-        return pd.read_csv("data/performance.csv", index_col=0)
-    else:
-        return pd.DataFrame()
+    if PERFORMANCE_CSV_PATH.exists():
+        return pd.read_csv(PERFORMANCE_CSV_PATH, index_col=0)
+    return pd.DataFrame()
 
 
 def performance():
@@ -23,7 +24,6 @@ def performance():
     valor_total = 100000
     volumes = [400000, 1000000]
 
-    # Carrega o estado anterior
     performance = carregar_estado_anterior()
     performance_final = pd.DataFrame()
 
@@ -40,33 +40,27 @@ def performance():
             sys.exit()
 
     print("Atualizando dados...")
-    requisicao = fd.get_resultado_raw()
+    requisicao = get_resultado_raw()
 
     for estrategia in estrategias:
         for ativos_na_carteira in combinacao_ativos_na_carteira:
             for vol_min in volumes:
                 df = requisicao.copy(deep=True)
 
-                # Remover ações financeiras (exceto WIZC3)
-                financeiras = set(fd.list_papel_setor(0)) - {"WIZC3"}
+                financeiras = set(list_papel_setor(0)) - {"WIZC3"}
                 df = df[~df.index.isin(financeiras)]
-
-                # Remover ADRs e BDRs
                 df = df[~df.index.str.contains("33|34")]
                 df = df[df["Mrg Ebit"] > 0]
                 df = df[df["Liq.2meses"] >= vol_min]
 
-                # Remover duplicatas com base nos 4 primeiros caracteres do ticker
                 df["First4Chars"] = df.index.str[:4]
                 max_values = df.groupby("First4Chars")["Liq.2meses"].transform("max")
                 df = df[df["Liq.2meses"] == max_values]
                 df.drop(columns="First4Chars", inplace=True)
 
-                # Calcular métricas
                 df["Earnings Yield"] = round(1 / df["EV/EBIT"] * 100, 1)
                 df["ROIC"] = round(df["ROIC"] * 100, 1)
 
-                # Ordenação
                 if estrategia == "Earnings Yield":
                     sorted_df = df.sort_values(
                         by=["EV/EBIT", "Liq.2meses"], ascending=[True, False]
@@ -78,12 +72,11 @@ def performance():
                     sorted_df = df.sort_values(
                         by=["Magic Formula", "Liq.2meses"], ascending=[True, False]
                     )
-                else:  # ROIC
+                else:
                     sorted_df = df.sort_values(
                         by=["ROIC", "Liq.2meses"], ascending=[False, False]
                     )
 
-                # Seleção dos ativos
                 sorted_df = sorted_df.head(ativos_na_carteira)
                 sorted_df["Quantidade"] = round(
                     (valor_total / ativos_na_carteira) / sorted_df["Cotação"], 0
@@ -96,7 +89,6 @@ def performance():
                         & (performance["Volume Mínimo"] == vol_min)
                         & (performance["Data"] == performance["Data"].max())
                     ].drop_duplicates()
-                    # Identificar entradas e saídas
                     ativos_atuais = set(sorted_df.index)
                     if not carteira_anterior.empty:
                         vendidos = carteira_anterior.loc[
@@ -106,18 +98,14 @@ def performance():
                             carteira_anterior.index.isin(ativos_atuais)
                         ]
                         comprados.drop(columns=["Cotação"], inplace=True)
-                        # calcula novo valor de ativos comprados
                         comprados = comprados.merge(
                             sorted_df[["Cotação"]],
                             left_index=True,
                             right_index=True,
                             how="left",
                         )
-                        comprados["Valor"] = (
-                            comprados["Cotação"] * comprados["Quantidade"]
-                        )
+                        comprados["Valor"] = comprados["Cotação"] * comprados["Quantidade"]
                         if len(vendidos) > 0:
-                            # calcula valor vendido de ativos vendidos
                             vendidos.drop(columns=["Cotação"], inplace=True)
                             vendidos = vendidos.merge(
                                 df[["Cotação"]],
@@ -125,19 +113,12 @@ def performance():
                                 right_index=True,
                                 how="left",
                             )
-                            vendidos["Valor"] = (
-                                vendidos["Cotação"] * vendidos["Quantidade"]
-                            )
+                            vendidos["Valor"] = vendidos["Cotação"] * vendidos["Quantidade"]
                             valor_vendido = vendidos["Valor"].sum()
-                            # Calcular quantidade de novos ativos comprados
                             ativos_mantidos = comprados.index
                             sorted_df["Quantidade"] = np.where(
                                 ~sorted_df.index.isin(ativos_mantidos),
-                                round(
-                                    (valor_vendido / len(vendidos))
-                                    / sorted_df["Cotação"],
-                                    0,
-                                ),
+                                round((valor_vendido / len(vendidos)) / sorted_df["Cotação"], 0),
                                 0,
                             )
                         sorted_df = sorted_df.merge(
@@ -152,30 +133,21 @@ def performance():
                         sorted_df["Quantidade"] = (
                             sorted_df["Quantidade_x"] + sorted_df["Quantidade_y"]
                         )
-                        sorted_df.drop(
-                            columns=["Quantidade_x", "Quantidade_y"],
-                            inplace=True,
-                        )
-                        sorted_df["Valor"] = (
-                            sorted_df["Quantidade"] * sorted_df["Cotação"]
-                        )
-                        valor_total_atualizado = sorted_df["Valor"].sum()
-                    else:
-                        valor_total_atualizado = valor_total
+                        sorted_df.drop(columns=["Quantidade_x", "Quantidade_y"], inplace=True)
                 sorted_df["Valor"] = sorted_df["Quantidade"] * sorted_df["Cotação"]
-
-                # Adicionar metadados
                 sorted_df["Estratégia"] = estrategia
                 sorted_df["Ativos na Carteira"] = ativos_na_carteira
                 sorted_df["Volume Mínimo"] = vol_min
                 sorted_df["Data"] = pd.to_datetime("today").date()
-
-                # Atualizar dataframe final
                 performance_final = pd.concat([performance_final, sorted_df])
 
-    # Salvar os resultados
-    performance_final.to_csv("data/performance.csv")
+    performance_final.to_csv(PERFORMANCE_CSV_PATH)
     print("Atualização concluída e salva.")
 
 
-performance()
+def main():
+    performance()
+
+
+if __name__ == "__main__":
+    main()
