+++ backend/services/performance_history_service.py
@@ -1,49 +1,18 @@
-import argparse
 import io
 import subprocess
 from pathlib import Path
 
 import pandas as pd
-from scripts.repair_performance_data import load_quote_matrix, repair_performance_dataframe
 
-"""
-run:
-poetry run python scripts/build_performance_history_parquet.py --since "2026-03-10 00:00:00"
-"""
+from backend.services.performance_repair_service import (
+    load_quote_matrix,
+    repair_performance_dataframe,
+)
+from shared.clients.parquet_client import write_parquet
+from shared.config import REPO_ROOT
 
 
-REPO_ROOT = Path(__file__).resolve().parents[1]
-DEFAULT_INPUT = Path("data/performance.csv")
 ROW_KEYS = ["Estratégia", "Volume Mínimo", "Ativos na Carteira", "Data", "papel"]
-
-
-def parse_args() -> argparse.Namespace:
-    parser = argparse.ArgumentParser(
-        description=(
-            "Gera um arquivo Parquet com todas as versoes commitadas de um CSV no Git."
-        )
-    )
-    parser.add_argument(
-        "--since",
-        default="2026-03-10 00:00:00",
-        help="Data inicial para filtrar commits no formato aceito pelo git log.",
-    )
-    parser.add_argument(
-        "--input",
-        default=str(DEFAULT_INPUT),
-        help="Caminho do CSV versionado dentro do repositorio.",
-    )
-    parser.add_argument(
-        "--output",
-        default=None,
-        help="Caminho de saida do parquet. Se omitido, usa data/<nome>_committed_since_<data>.parquet.",
-    )
-    parser.add_argument(
-        "--include-working-tree",
-        action="store_true",
-        help="Inclui a versao atual do CSV no workspace, mesmo antes de commit.",
-    )
-    return parser.parse_args()
 
 
 def run_git_command(*args: str) -> str:
@@ -95,17 +64,14 @@ def load_working_tree_csv_version(input_path: Path) -> pd.DataFrame:
     return frame
 
 
-def main() -> None:
-    args = parse_args()
-    input_path = Path(args.input)
-    output_path = (
-        Path(args.output)
-        if args.output is not None
-        else build_default_output_path(input_path, args.since)
-    )
-
-    history = load_committed_csv_versions(input_path, args.since)
-    if args.include_working_tree:
+def build_performance_history(
+    input_path: Path,
+    output_path: Path,
+    since: str,
+    include_working_tree: bool,
+) -> pd.DataFrame:
+    history = load_committed_csv_versions(input_path, since)
+    if include_working_tree:
         history = pd.concat(
             [history, load_working_tree_csv_version(input_path)],
             ignore_index=True,
@@ -127,14 +93,8 @@ def main() -> None:
         .drop(columns="_commit_sort")
         .reset_index(drop=True)
     )
+
     destination = REPO_ROOT / output_path
     destination.parent.mkdir(parents=True, exist_ok=True)
-    history.to_parquet(destination, index=False, engine="pyarrow")
-
-    print(f"Arquivo gerado: {destination}")
-    print(f"Linhas: {len(history)}")
-    print(f"Commits: {history['commit_hash'].nunique()}")
-
-
-if __name__ == "__main__":
-    main()
+    write_parquet(history, destination, engine="pyarrow")
+    return history
