+++ shared/__init__.py
@@ -1,0 +1,1 @@
+"""Shared modules used by app and backend."""
