+++ shared/clients/__init__.py
@@ -1,0 +1,1 @@
+"""Shared clients for external systems and storage."""
