+++ shared/clients/fundamentus_client.py
