+++ shared/clients/yahoo_benchmark_client.py
